+++ tests/ops/test_audio.py
@@ -1,7 +1,7 @@
 """Tests for ``ai.ops.audio`` dispatch behavior.
 
 Gateway wire-format specifics live in
-``tests/providers/ai_gateway/test_generate_audio.py``; these tests cover
+``tests/providers/ai_gateway/test_protocol.py``; these tests cover
 the ops layer: parameter defaults and provider dispatch.
 """
 
@@ -13,14 +13,13 @@ from ai import ops
 from ..providers.ai_gateway.conftest import user_msg
 
 
-class TestUnsupportedProvider:
-    async def test_generate_audio_raises_not_implemented(self) -> None:
-        provider = ai.get_provider("openai", api_key="sk-test")
-        model = ai.Model(id="tts-1", provider=provider)
+async def test_generate_audio_raises_not_implemented() -> None:
+    provider = ai.get_provider("openai", api_key="sk-test")
+    model = ai.Model(id="tts-1", provider=provider)
 
-        try:
-            await ops.generate_audio(model, [user_msg("Hello!")])
-        except NotImplementedError as exc:
-            assert "generate_audio" in str(exc)
-        else:
-            raise AssertionError("expected NotImplementedError")
+    try:
+        await ops.generate_audio(model, [user_msg("Hello!")])
+    except NotImplementedError as exc:
+        assert "generate_audio" in str(exc)
+    else:
+        raise AssertionError("expected NotImplementedError")
